// assets/script/protobuf/holdem/req_enter_room_pb.d.ts
export class ClientMessageEnterRoom extends jspb.Message {
  hasRoom(): boolean;
  clearRoom(): void;
  getRoom(): protobuf_holdem_define_pb.Room | undefined;
  setRoom(value?: protobuf_holdem_define_pb.Room): void;

  hasGps(): boolean;
  clearGps(): void;
  getGps(): protobuf_holdem_define_pb.GPS | undefined;
  setGps(value?: protobuf_holdem_define_pb.GPS): void;

  getMttPartialBringIn(): number;
  setMttPartialBringIn(value: number): void;

  getObserver(): boolean;
  setObserver(value: boolean): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): ClientMessageEnterRoom.AsObject;
  static toObject(includeInstance: boolean, msg: ClientMessageEnterRoom): ClientMessageEnterRoom.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: ClientMessageEnterRoom, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): ClientMessageEnterRoom;
  static deserializeBinaryFromReader(message: ClientMessageEnterRoom, reader: jspb.BinaryReader): ClientMessageEnterRoom;
}

export namespace ClientMessageEnterRoom {
  export type AsObject = {
    room?: protobuf_holdem_define_pb.Room.AsObject,
    gps?: protobuf_holdem_define_pb.GPS.AsObject,
    mttPartialBringIn: number,
    observer: boolean,
  }
}